Move base settings deeper into module tree.
Similar to "manage.py", "settings.py" is far too generic a module name to be
included in a library. Move that module deeper in the directory tree.
