+++ Dockerfile
@@ -2,7 +2,7 @@ FROM python:3.5-alpine
 
 COPY ["regparser", "/app/src/regparser/"]
 COPY ["interpparser", "/app/src/interpparser/"]
-COPY ["settings.py", "requirements.txt", "setup.py", "manage.py", "/app/src/"]
+COPY ["requirements.txt", "setup.py", "manage.py", "/app/src/"]
 VOLUME ["/app/cache", "/app/output"]
 
 WORKDIR /app/src/
