+++ regparser/history/annual.py
@@ -8,9 +8,9 @@ from datetime import date
 import requests
 from cached_property import cached_property
 
-import settings
 from regparser.index.http_cache import http_client
 from regparser.tree.xml_parser.xml_wrapper import XMLWrapper
+from regparser.web.settings import parser as settings
 
 CFR_BULK_URL = ("https://www.gpo.gov/fdsys/bulkdata/CFR/{year}/title-{title}/"
                 "CFR-{year}-title{title}-vol{volume}.xml")
