+++ regparser/layer/terms.py
@@ -6,12 +6,12 @@ from collections import defaultdict, namedtuple
 
 import inflection
 
-import settings
 from regparser.layer import def_finders
 from regparser.layer.layer import Layer
 from regparser.layer.scope_finder import ScopeFinder
 from regparser.tree import struct
 from regparser.tree.priority_stack import PriorityStack
+from regparser.web.settings import parser as settings
 
 try:
     key = ('(?i)(p)erson$', '\\1eople')
