+++ regparser/regs_gov.py
@@ -2,7 +2,7 @@ import logging
 from collections import namedtuple
 
 from regparser.index.http_cache import http_client
-from settings import REGS_GOV_API, REGS_GOV_KEY
+from regparser.web.settings.parser import REGS_GOV_API, REGS_GOV_KEY
 
 REGS_GOV_DOC_API = REGS_GOV_API + 'documents.json'
