+++ tests/test_web_api.py
@@ -10,11 +10,11 @@ from mock import Mock, patch
 from rest_framework.test import APITestCase
 from six.moves.urllib.parse import urlparse
 
-import settings
 from regparser.web.jobs.models import job_status_values
 from regparser.web.jobs.utils import (create_status_url, eregs_site_api_url,
                                       file_url)
 from regparser.web.jobs.views import FileUploadView as PatchedFileUploadView
+from regparser.web.settings import parser as settings
 
 fake_pipeline_id = uuid4()
